Cleanup of security scanning services
