+++ UnitystationLauncher/Models/ContentScanning/ScanningTypes/MResScope.cs
@@ -1,0 +1,5 @@
+namespace UnitystationLauncher.Models.ContentScanning.ScanningTypes;
+
+public record MResScope
+{
+}
