+++ UnitystationLauncher/Models/ContentScanning/ScanningTypes/MResScopeAssembly.cs
@@ -1,0 +1,9 @@
+namespace UnitystationLauncher.Models.ContentScanning.ScanningTypes;
+
+internal sealed record MResScopeAssembly(string Name) : MResScope
+{
+    public override string ToString()
+    {
+        return $"[{Name}]";
+    }
+}
