+++ UnitystationLauncher/Models/ContentScanning/ScanningTypes/MResScopeType.cs
@@ -1,0 +1,10 @@
+namespace UnitystationLauncher.Models.ContentScanning.ScanningTypes;
+
+
+internal sealed record MResScopeType(MType Type) : MResScope
+{
+    public override string ToString()
+    {
+        return $"{Type}/";
+    }
+}
