+++ UnitystationLauncher/Models/Enums/ClientRequest.cs
@@ -1,0 +1,8 @@
+namespace UnitystationLauncher.Models.Enums;
+
+internal enum ClientRequest
+{
+    URL = 1,
+    API_URL = 2,
+    Host_Trust_Mode = 3,
+}
