+++ UnitystationLauncher/Services/Interface/ICodeScanConfigService.cs
@@ -1,12 +1,15 @@
 ﻿using System.Threading.Tasks;
+using UnitystationLauncher.Models.ConfigFile;
 
 namespace UnitystationLauncher.Services.Interface;
 
-public interface IGoodFileService
+public interface ICodeScanConfigService
 {
     public Task<(string, bool)> GetGoodFileVersion(string version);
 
     public Task<bool> ValidGoodFilesVersion(string goodFileVersion);
 
     public string SanitiseStringPath(string inString);
+
+    public Task<SandboxConfig> LoadConfigAsync();
 }
