--- UnitystationLauncher/Services/Interface/IFileService.cs
@@ -1,10 +1,0 @@
-﻿using System.IO;
-
-namespace UnitystationLauncher.Services.Interface;
-
-public interface IFileService
-{
-    public StreamReader OpenText(string path);
-
-    public bool Exists(string path);
-}
