+++ UnitystationLauncher/Services/Interface/IGameCommunicationPipeService.cs
@@ -1,0 +1,4 @@
+public interface IGameCommunicationPipeService
+{
+    public void Init();
+}
